> Restores specialRFact to 150% (i.e. 1.5). Since the value of 2.5 is introducing leader-casual agent global communication; which isn't the desired scenario of distributed local swarm comuncation
> Average numATgoal is computed every 50 iterations (iterWindow), starting from the
iteration in which the first casual agents enters the goal area. It CAN'T be computed
in each iteration (as done before) beacause it over penalizes agents that arrive just a bit latter.
Since from iteration-to-iteration agents dont move much, its better to leave a iteration window to give them
enough time to move